fix(nlp): Fixes paths that were not automatically renamed.
