+++ archai/nlp/datasets/corpus.py
@@ -8,10 +8,10 @@ import numpy as np
 import torch
 
 from archai.common import utils
-from archai.nlp.tokenizer_utils.vocab_base import VocabBase
-from archai.nlp.tokenizer_utils.word_vocab import WordVocab
-from archai.nlp.tokenizer_utils.bbpe_vocab import BbpeVocab
-from archai.nlp.tokenizer_utils.gpt2_vocab import Gpt2Vocab
+from archai.nlp.datasets.tokenizer_utils.vocab_base import VocabBase
+from archai.nlp.datasets.tokenizer_utils.word_vocab import WordVocab
+from archai.nlp.datasets.tokenizer_utils.bbpe_vocab import BbpeVocab
+from archai.nlp.datasets.tokenizer_utils.gpt2_vocab import Gpt2Vocab
 
 from archai.nlp.datasets.lm_iterators import LMMultiFileIterator, LMOrderedIterator, LMShuffledIterator
