+++ archai/nlp/datasets/tokenizer_utils/bbpe_vocab.py
@@ -10,11 +10,11 @@ from overrides import overrides
 from tokenizers import ByteLevelBPETokenizer
 from transformers import PreTrainedTokenizerFast, PreTrainedTokenizer
 
-from archai.nlp.models import nvidia_utils as nv_utils
-from archai.nlp.tokenizer_utils.vocab_base import VocabBase
-from archai.nlp.tokenizer_utils.token_config import TokenConfig
+from archai.nlp.datasets import distributed_utils as nv_utils
+from archai.nlp.datasets.tokenizer_utils.vocab_base import VocabBase
+from archai.nlp.datasets.tokenizer_utils.token_config import TokenConfig
 from archai.common import utils, common
-from archai.nlp.tokenizer_utils.special_token_enum import SpecialTokenEnum
+from archai.nlp.datasets.tokenizer_utils.special_token_enum import SpecialTokenEnum
 
 class BbpeVocab(VocabBase):
     def __init__(self, save_path:str, vocab_size:int, pad_vocab_size=False,
