+++ archai/nlp/datasets/tokenizer_utils/vocab_base.py
@@ -7,7 +7,7 @@ from collections import abc
 from overrides import overrides, EnforceOverrides
 import torch
 
-from archai.nlp.tokenizer_utils.special_token_enum import SpecialTokenEnum
+from archai.nlp.datasets.tokenizer_utils.special_token_enum import SpecialTokenEnum
 
 class VocabBase(EnforceOverrides, abc.Sized):
     @abstractmethod
