+++ archai/nlp/models/hf_gpt2.py
@@ -8,7 +8,7 @@ import torch.nn.functional as F
 from transformers import CONFIG_MAPPING, AutoModelForCausalLM
 
 from archai.nlp.models.archai_model import ArchaiModel
-from archai.nlp.models.models.model_utils import map_to_list
+from archai.nlp.models.model_utils.utils import map_to_list
 
 
 class HfGPT2(ArchaiModel):
