+++ archai/nlp/scoring_metrics/vocab_wrapper.py
@@ -5,8 +5,8 @@ from typing import List, Tuple
 import numpy as np
 
 from archai.common.lru_cache import LRUCache
-from archai.nlp.scoring.scoring_utils import WORD_TOKEN_SEPARATOR_SET, RE_SPLIT
-from archai.nlp.tokenizer_utils.vocab_base import VocabBase
+from archai.nlp.scoring_metrics.scoring_utils import WORD_TOKEN_SEPARATOR_SET, RE_SPLIT
+from archai.nlp.datasets.tokenizer_utils.vocab_base import VocabBase
 
 TOKENIZER_FILTER_TOKEN_IDS_CACHE_SIZE = 65536
